Align new rep current segment with old for live streams
Previous behavior with live streams was to always start a new stream ~12 seconds from the live edge (except multiperiod - only last period starts this way). It does this by selecting `current_segment_` when starting the stream. However we now need to keep the same position for `current_segment_`  when reopening a stream that has switched reps as the new rep will not have or at least an up-to-date `current_segment_`.