Fix tests
The list of downloaded segment urls is now kept within the stream itself.
This list also includes the initialization segment so tests are updated accordingly.
The amount of segments that the worker thread downloads is not reliably predictable. Sometimes it will get only the amount of segments that we ask to 'read' but other times it will run away and start filling the available buffers on top of the reads requested, making the previous use of `downloadedUrls.back()` unreliable.